server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic-settings==2.6.0
numpy==1.26.4
scipy==1.14.1
soundfile==0.12.1
funasr==1.3.1
modelscope==1.18.1
webrtcvad-wheels==2.0.14
